rework backend of software to include intertrial intervals
This commit changes a lot of the backend structure. It allows to specify
the time for an intertrial interval (break between stimulus). Also
the description label is used as an info box and all other use-cases
are removed. Saving of data is outsourced to an external thread to keep
the GUI responsive. Included a start-phase for the subject to get ready.